Split lock screen interactor into entry and view interactors
