+++ padlock/src/main/java/com/pyamsoft/padlock/lock/LockScreenActivity.java
@@ -151,9 +151,6 @@ public class LockScreenActivity extends ActivityBase {
     PreferenceManager.setDefaultValues(getApplicationContext(), R.xml.preferences, false);
 
     Injector.get().provideComponent().plusLockScreenComponent().inject(this);
-    if (savedInstanceState == null) {
-      presenter.resetFailCount();
-    }
 
     populateIgnoreTimes();
     getValuesFromBundle();
